Add $ check for minion attacks
If a player has an evil twin cast the evil twin will have a master but
will not be on the player's minion list. This is expected behavior of
evil twins because they usually attack only the initial target. This
commit adds a $ check for the minion list to prevent errors if there is
no list.