Send self to SomeoneSaid
Message that was supposed to be displayed on deletion was not being sent
with the required parameters to display it. Caused errors on some
screens (and of course didn't display the message). Now sends what=self
so when range is checked further up, the message can be sent to the
required objects.