Armor of Gort damage cap bug fix
Armor of Gort wasn't capping damage taken properly, because order
mattered. To receive the full benefit of Armor of Gort, a player would
have to buff while already wearing his armor, and then never switch to
any other gear. Switching would lose the benefits due to pushing Armor
of Gort later in the list of defense mods.

This change moves the capping mechanism before all defense modifiers.

This change will make Armor of Gort appear stronger.
In reality, it will just be working as intended, and as it always did for
a select few players with bad buffing habits.

There was a myth that casting Gort in armor made it better. Turns out,
in some fashion, it was true. This was why.

Example: a 30 damage hit with Gort first would be capped to 15,
then reduced by plate and jonas shield by up to 10, to 5 damage taken.

A 30 damage hit with Gort last (if cast in robes then switch to plate)
would have plate and jonas shield go first, reducing 30 to possibly 20,
then capped to 15. Thus the gear-switching changed your minimum
damage taken from 5 to 15. It's a pretty big swing in many common situations,
and there's no way for players to know how badly this was tilting things.