Fix SafePlayerAttack errors on some kills.
Currently the killer's room is always used to determine whether they
receive a penalty or not. There are cases where the killer will be
offline or out of the room (firewall, minions) when the kill occurs, so
the victim's room should really be used (guaranteed to be at the killed
location, and online).