Add viVisionDistance variable.
Added a new class variable/property for monster vision -
viVisionDistance (and piVisionDistance). Splits the function of vision
away from viDifficulty. The number represents the radius in which a
monster can 'see', in equivalent row or col units.

Default vision is 10, completed special cases of revenant and living
statue (rest to come in separate commit).