Refactor vision check in brain.kod.
Now checks every time GetHatred is called (but not if the mob itself was
attacked). Returns 0 and doesn't switch to the target if outside vision
range.

Monsters attempt to target switch to the player on 20% of player moves
rather than 10%, so they aren't standing around so much.

piChaser_factor lowered to 100; if a player runs into a monster that has
very high hatred for them it may aggro even if another mob is chasing
the player.

Reflection vision raised to 60 as low vision was stopping them from
working correctly.