Fix bugs in SetMaster target handling.
The ResetMonsterChasers message called by entering a room was masking
bugs caused by SetMaster adding and removing targets but not correctly
handling the piMonsterChasers property in Player used to keep track of
how many mobs are chasing a player. This should now work correctly.
Also modified SomethingMoved in brain.kod to take the correct action
when the mob is in STATE_LIMBO rather than just print a debug string.