Post setting of poOriginal to $ to avoid errors.
On rare instances, someone looking at a reflection or evil twin while it
is in the process of being deleted will cause errors to be thrown as it
tries to send messages to $. This can occur with other messages that
call ToCliObject in user.kod. The removal of poOriginal will now be
posted, so any calls that require this data will still succeed while in
the process of deletion.