SetMaster refactor, behavior flag check for SomethingLeft.
Much simpler to handle the special ET case by having a SetMaster in
eviltwin.kod, same for reflections. Added a behavior flag check (for
AI_MOVE_FOLLOW_MASTER) in SomethingLeft so minions don't follow if
they aren't supposed to (i.e. ET). Also, stop minions in general
from attacking revenants (these are a punishment for the master,
not for the minions.)