Namecolor flag addition - RGB value.
Split name color flags off from object flags to its own field. This now
sends the RGB value for the desired color. The client seems to not draw
the RGB value, instead using the closest match, however the Ogre client
should be able to use this info to pick the proper RGB value.