Allow UpdateTimeValues to reset piTimeAttackedPlayer.
This is the bug causing players to be unable to rescue/elusion. The
piTimeAttackedPlayer property needs to be set to current time with other
properties in UpdateTimeValues so when the system time is larger than
the KOD integer limit and is reset, it can be set to the new lower time
value.

Running send c player UpdateTimeValues after this code is live will fix
the bug for all players (after the requisite time has passed).