Lower vigor drain to previous levels.
This should match the vigor drain on 101/102. move.c used to send 18 for
walk speed, 36 for run but has been updated to send 25/50. This was then
modified to 5/6ths that value for vigor drain, presumably to match the
behavior of an earlier version of walk/run speed. We now multiply by 3/5
to match that value.