German menu fixes, string length increase.
Increased MAXSTRINGLEN to 512 to properly display the German strings.
I'm not actually sure why this limit was set so low; I had already
bumped it up to 255 from 200 to fix bio crashes a long time ago. I've
been testing everything associated with it and can't seem to find any
issues with a larger number here.