Drop vigor drain from narthyl worms.
Vigor drain caused by narthyl worm hits was far too high with safe spots
removed and the possibility of multiple aggro. Lowered the vigor loss
min bound from 3 to 0, and the max from 7 to 3. In testing vigor still
noticeably drops, but no longer requires the player to rest or be
constantly too full to eat.