Allow monsters to have minions. Convert Seduce and Animate for use by monsters.
Moved bulk of the minion control code from Player to Battler. Player and
Monster both have their own version of CommandMinionAttack due to the
extra checks needed for players. Minions for monsters function the same
as they do for players and recursion is also allowed - players can
theoretically have minions that have minions themselves.

Seduce and Animate have been modified so that monsters can also cast
them if necessary. If a monster casts Seduce, it is allowed to charm the
target.