Vault change
The vault patch from 103, but reduced the number of items to 150 instead
of their 600 which might be a little overkill. 150 might be to much too
but I leave that to the developers to decide.

The vaults allow you to carry 150 items (stacks of reagents/food count
as one item).
People are building and gathering loot but have to store it on multiple
mules, this would reduce the number of mules people are using.