Add string handling to SendMessage and Postmessage.
Modified C_SendMessage and C_PostMessage so that they can be passed a
string (TAG_STRING) containing a message name, and if that name is valid
can convert it to a message ID and use that to process the message. This
code is only executed if the sent value fails an existing check (the
TAG_MESSAGE check) so existing calls to both these functions should not
take any longer to run.