Add message and debug string handling to C_SetString.
C_SetString will now handle adding debug strings (i.e. " " strings in
KOD) and messages (i.e. @StartEvent) to a string. Debug strings have the
contents added as-is, messages have the message name added.

C_SetString will also now create a string if it is passed $ as the
string to set to, so that a string can be created and set with one call
(i.e. s = SetString($,"string") rather than two.

Added a commend preceding error message printed when GetTagData
encounters TAG_MESSAGE as message saving *is* supported, however not
advised. Added the ability for GetDataNum to return a valid message ID -
this is used by admin functions only, and future improvements to message
handling will require this change.