Add checking for duplicate language strings.
make_resource now adds its (new) completed resources to the resource
list for the current class, and when each new resource is created it is
first checked against this list. If a duplicate resource ID is
encountered, the appropriate language ID (resource array position) is
checked and if a string with the same language ID is already present, an
error is thrown. New resources are returned to the parser, ones that
existed already are not.