Add SwapListElem C function, swaps the data in two list elements.
Takes a list and two ints and swaps the data at those elements in the
given list. Returns NIL (same as SetNth/SetFirst).

Added some info regarding ConsEnd and SwapListElem in ccode.c, and also
modified ConsEnd in list.c to print an error if passed a list with more
than 500 elements (as it has to traverse the list to append a list node
to the end, large lists can potentially pause blakserv).