Allow C_ConsEnd and C_AddListElem to handle $ lists.
Also added error checking to C_Cons so that it doesn't try to add an
element to something that isn't either $ or a valid list. Doing so
currently 'works', but the list structure returned is not valid. On my
computer this check adds ~2ns per call of Cons, but allows errors when
using Cons to be detected much faster.