Stop mouselook and tab repeat toggle on key held down.
This change causes mouselook toggle and tab keys to be treated as repeat
actions, so both can be repeated instantly. It also stops the idle
keyboard polling from performing either of those actions, the net result
being that mouselook/tab window can always be toggled, but will not
toggle repeatedly if the key is held down. Client min version
incremented.