Allow potion/scroll use during attack cooldown.
Discussed in IRC - currently the default behavior for using an item
during attack cooldown is to have it fail with no extra info, and remove
1 from piHits. This change prevents wands from being used during the
cooldown, but allows potions and scrolls.

Perhaps it isn't realistic to drink a potion or read a scroll during
attack cooldown, but the presented argument was that it would open up
some strategy with regard to fast item use (while still preventing
double attacks using lightning or other offensive wands).