Waylay Oasis
Waylay Oasis is one of the few safe places in the desert areas. It's
inside the narrow trails of the cliff plateau, and it has several unique
features. One of them is a type of guild hall we've never seen before,
which is present directly on the same map, hidden among the cliffs. It's
quite possible for a guild that owns the Waylay Oasis hall to launch
attacks on those who think they are safe below.

Work in progress