Check lights against player view.
Checks the bounding box of each poly against the player's view before
adding lighting effects to it. Doesn't apply to static lightmaps
(generated elsewhere, and only when the room is initially loaded).

The purpose of this is to provide some speedup to drawing large amounts
of dynamic lights - while it doesn't prevent the massive slowdown that
occurs when a bunch of them are in view, it does drop light rendering
time to zero if the player isn't looking at them (and if the player can
only see a subset of the illuminated polys, rendering is faster).