Refactor D3D lighting code.
Instead of traversing the BSP tree once and checking each light against
each vertex of each floor, ceiling and wall, the renderer now traverses
the tree for every light.

If the light isn't present in the bounding box of the node, traversal
stops at that node and the other branches of the tree are checked. This
is exponentially? faster than how it was being done.

Fixed a compiler warning in graphics.c.