Add torch template to room editor.
Places a torch at the selected (by mouse) x, y coordinates. User can
select whether to place the torch going north, south, east or west. If
placed within 4 units of a linedef, the torch will attach to that
linedef at the closest point (shifting the torch appropriately, not
moving the linedef). When placed, a message will display giving the user
the coordinates to place the dynamic light for this torch at.