Copy slope data when copying sectors, fix undo behavior.
Copying sectors with slopes will now also copy the slope data correctly.
To facilitate this, copied vertexes keep track of the new vertex and
when slope data is copied, the old vertex is replaced by the new one in
the new slope.

Max number of undos increased from 5 to 50, and this list is no longer
cleared on saving a level.