Add correct sloped floor/ceiling heights to roo grid heightmap.
Added GetFloorHeight and GetCeilingHeight functions which return the
height along a sloped surface correctly for a given x, y coordinate.
Values divided by 16 to match leaf->floor_height/ceiling_height. This
fix is necessary as we use the height maps for determining whether
attack range is valid or not.

Changed tabs to spaces in move.cpp.

Updated binaries.