Add postbuild.bat to build, update postbuild.
postbuild.bat will now check the patcher's settings file in default
location to obtain the client resources. It checks each meridian client
location for the files, and will stop when it finds a valid one.

It also now creates a shortcut in ..\run\localclient to meridian.exe
which can connect to localhost.