Fix WA_SPELLCASTER not spawning all spells.
Also changed them to use the SID rather than the spell object, as the
object can be deleted/recreated. Added a message to weapon.kod to
replace the spell object in all existing spellblades with the
appropriate SID.

Removed the swap and invisibility spells from the spellblade spawn list
(no current weapons actually have them).

Added the spell cast by the weapon to the weapon's description.