Stop players and mobs dropping loot inside walls.
This fix creates the dead body and monster loot exactly where the
monster was standing, instead of using the default value for that
finerow/finecol (32, 32) which sometimes placed the loot inside walls.

Players will also drop their loot exactly where they are standing
when killed.

The ReqSomethingMoved checks before item drops have all been removed -
if the player or monster can stand in that space, we will assume the
loot can be placed there also. The alternative is having loot rarely
still not drop if the player or monster is standing too close to
a wall.