Fix rentable room setup bugs
Occasionally there were issues with different wallpapers being added to
the same room on initial room setup, due to differing default values for
wallpaper in the .roo and .kod files.

A similar issue was occurring with the bed and pillows not being added
correctly, if the randomly chosen bed sector was the default one.