Add rentable rooms to Ko'catan Inn.
Graphics and draft copy of the code by Diggie. Adds a new rentable room
type to the Ko'catan Inn, with one wall graphic, and uses existing
ceiling and floor graphics. New ceiling and wall types can be chosen in
any rentable room.

Simplified the string that is assembled when a player asks about room
cost in all 3 Innkeepers (this is now a resource with 3 parameters).