Fix vault and weight/bulk bugs.
Some bad logic was causing players to be able to carry over their
maximum weight/bulk limit, and the bad return values from the
weight/bulk calculations was causing errors in several places. Added
bounds and checks in Holder and NumberItem respectively to prevent this
happening (redundant when called together, but that might not always be
the case).

Removed an unnecessary debug message from VaultWithdraw in Monster that
was being triggered by mundane behavior. Added a check for valid
number_list parameter before using it in Buy, as this was ending up $
for a reason I can't find. I think it may be a client sending the wrong
info here rather than anything in the server code.

Formatted storage.kod, and added a check for GetNumberCanHold returning
less than 0 to Storage (which shouldn't happen with the above fixes, but
just to be safe...). Added a check for immortal DMs so they can withdraw
items from the vault without the server throwing errors and having it
fail. Improved a debug message that alerted me to the vault issues in
the first place.