Add spellpower to enchantments.
Added an extra element to enchantments stored in the plEnchantments
list, storing the spellpower the spell was cast at. Modified all list
length checks to use constants in case further modification of the
enchantments is ever needed.

Standardised a few spells that had extra effects done in CastSpell, as
these should be done in GetStateValue. The reasoning for this is that
GetStateValue is called once when the enchantment is started, so if
anything needs to apply the enchantment via StartEnchantment, the extra
spell effects will always be applied correctly.