QuestEngine refactor, quest tab in client, new string formatters.
Added %r resource string formatter, allows a resource with its own
string formatters to be added in place of the %r (processes the resource
completely) before moving further in the resource containing the %r.

Added numbered parameters for resource string formatters, so a resource
with "You hit the %s$1 with your %s$2." can be displayed with the
parameters in a different order in another language resource without
changing the order sent by the server. In this case, "You use your %s$2
to hit the %s$1." would display the correct information.

Refactored QuestEngine to use quest template objects. Added the first
one (Shal'ille disciple quest). Added quest tab to the client, uses
the quest template objects to display information about the quest
to the user.