Fix parameter handling for admin commands.
When reading resource parameters, the code was assuming that the
parameters were present without actually checking for them. This was
most noticeable with the "lock" command, which displayed garbage if no
lock text was given as it reads past the end of the "lock" string to
find the requested lock text.

Changed the 'R' parameter-based admin commands to use strtok correctly,
so the commands will now fail if the parameter is not present. This does
mean that lock will no longer work on its own, the alternative is to let
NULL through here but handle it correctly for every single
resource-based admin command (most of them cause blakserv to crash when
passed NULL).

Added a few safety checks.