CheckServerMessage message length handling changes.
Added error checking to %r formatter, now calls "missing resource" box
if there's a length error. CheckServerMessage also now accepts a pointer
to len (length of server message) so the remaining length is kept
accurate if len is used after/for the recursive call.

Added text to the missing resource dialog telling user to contact an
administrator if the message continues to appear (if the client is
updated, the presence of this dialog box indicates a bug).