Handle transition from 3-element enchantment to 4-element.
While we can make sure nobody has enchantments on during the update
process, other groups using this codebase might not be following closely
enough to anticipate the potential issues this can cause. These changes
should mitigate the effects if enchantments are left on players during
the update, but should be removed in the following update.