Fix %r resource formatter message handling.
Resources that contain %s, text and one or more following %r formatters
were not adding the text part to the message displayed to the user. The
fix was to copy in tempfmt (message before formatter/between formatters,
including the 'r'), followed by removing the 'r' (% is already ignored)
and after that adding in the part of the message added in place of the
%r formatter. This is necessary for more complex uses of %r such as
spell descriptions.