Allow $number string parameters to work with %r.
The code responsible for checking and reordering parameter fields using
$1, $2 etc. has been moved to a separate function (CheckMessageOrder).
This function can recurse on hitting an %r parameter and will reorder
its contents if necessary.

This function should be called before CheckServerMessage, but only
really needs to be done for messages that could possibly contain the
$number formatters (so no need to handle mail and news, but have to
handle all description/string messages).