Stop modified guild shields causing errors.
Admin modified guild shields cause errors to be logged on the server,
and also "missing resource" boxes for the holder of the shield. The
shield will now send the correct data (i.e. nothing) if
prShield_emblem_drop is modified to be $. The shield now also displays
correctly in the inventory.