Add variable NPC response quest prizetype.
Allows random destination NPCs to be able to give individual quest
success responses. QuestTemplate handles the dialog to keep it
centralised, rather than spread it out over the NPC kod files.
Moved the Library speech lib responses for disciple quests into
the QuestTemplate file for those quests, and reverted the change
to Monster that allowed both QuestEngine and Library responses to be
given for the same string.