Necromancer scenario/guild/quest fixes.
The way the necromancer guild was being handled was incorrect - the Lich
was keeping track of it, and the guild was being deleted on RecreateAll.
NecromancerBalance now keeps track of the guild.

Added a new quest restriction - Q_R2_NECRO_GUILD_ACTIVE. The necro guild
quest requires Q_R2_NECRO_GUILD_ACTIVE to be FALSE to pass restriction,
as the Lich is the destination NPC for the quest when she is active
instead of the DeadLich. This also means the scheduling chance doesn't
have to be modified (the amulets and the guild's existence will
determine whether the quest can be given) and if something does go wrong
and a Lich is left in the hall instead of a DeadLich, players could
still create a guild, if one does not already exist.

Necromancer amulets now correctly handle the guild disbanding, and kill
off the guild members.