Add gain chance and TP quest prizetypes.
Added prizetypes to award training points and gain chance. Moved the
"roll for tougher" code out of AdvancementCheck in Player to its own
message, RollForTougher, that takes the monster level as a parameter.
Also added an AddGainChance message, that can add (or remove) gain
chance. Added the TP and gain chance prizetypes to the Moxal scam
quests, mostly as examples of their implementation (100TP for one
quest, 10 gain chance for the other). Perhaps worth doing, but not
worth farming.