Add message to fix players when changing spell level.
FixSpellLevelChange will remove a spell from players and readd it if
they can learn it. Can be used when changing a spell level to make sure
only players who can learn the new level of the spell will have it. If
the spell isn't readded, the player is given the equivalent number of
training points for the % they had in the spell.

Added some code to the training point system in Player to get around the
1000 point cap if necessary.