Add $0 number parameter for resource inline variables.
If an inline variable (%s, %i, %d, %r, %q) is given with a number
parameter of $0 (e.g. You hit the %s$0%s., the $0 being for the first %s
here) that parameter will not be added to the displayed message. This is
necessary if an added language doesn't require all the information (e.g.
the definite or indefinite article) for a resource and needs to make
sure it doesn't get displayed. Works with %r/%q/random $number orderings
etc.