Add timing changes, clean up admin debug info.
Added a config boolean to turn the timing of C calls and kod messages on
and off. Currently, using QueryPerformanceCounter to record time on
Amazon-based servers does not work correctly (takes a long time to get
the time) so this is off by default. The checks for the timing use
kod_stat.frequency to determine whether to run or not, as this is faster
than checking the config setting. Toggling the config setting via admin
window/server interface will set this value and turn the system on.

Removed a config int for BLAKOD_MAX_STATEMENTS, turned into a define
instead. This isn't something we need to set in config.

Commented out the check of DEBUG_INITLOCALS in InterpretAtMessage (same
function as the previous one). Again this isn't something we need to
really use, but have left the code intact in case someone ever does want
to use it.

Added a GetMicroCountDouble function to blakserv that returns the number
of microseconds in QueryPerformanceCounter. Even with the overhead on this
function on the Amazon server, we need this metric to keep track of time
spent interpreting (was already used but at the millisecond-precision level
when we can do better). The main purpose for switching to microseconds was
that the time spent interpreting was very inaccurate, as any call less than
1 millisecond (i.e. most of them) was not being recorded.