Add TAG_TABLE for blakod table datatype, add calloc infrastructure.
Added TAG_TABLE data type (tag number 13), moved TAG_OVERRIDE to 14.

Added AllocateMemoryCalloc macro which keeps track of memory allocated
by calloc, to more quickly allocate arrays with NULL structs as opposed
to malloc/for loops.

Converted the storage for blakod tables from a linked list into a
dynamically sized array, similar to blakod lists. This allows O(1)
lookup time, which will be important for large numbers of tables.
Memory for table structs is allocated on startup, and only the memory
for the hashes themselves is allocated while the server is in operation.

Added a ForEachTable function for future use in saving/loading tables.