Clean up built-in acct/obj code, fail old kodbase.txt builds.
Simplified the built-in acct/object code, making it a bit clearer what
is being done.

Added a check for new built-in objects to the kodbase loading code in
compiler, which will fail the build if an incompatible kodbase.txt is
loaded. Without this check, the build would succeed and the server would
be unable to use the built-in object.