Add IsListMatch() C function, clean up qormas.kod.
Added IsListMatch(list1, list2) C function that takes two lists and
returns TRUE if lists are identical, FALSE otherwise. Replaced IP checks
in System and Qormas with this new function, as with the change to IPv6
checking two IPs requires up to 32 Nth calls if the IPs match.
IsListMatch() is roughly as fast as 4-5 Nth calls when comparing two
16-element lists (and much easier to write).

Cleaned up whitespace in qormas.kod, and simplified a few statements.