Improve save game/load game built-in object handling.
Added a version to savegames. Old savegames have no version and are
version 0, new savegames are version 1. Version 0 savegames only load a
System built-in object, version 1 will load as many as are specified.
This versioning allows unlimited built-in objects to be added in the
future without exhausting the available SAVE_GAME enum space (and also
simplifies the process of adding built-ins).