Prevent stack overflow with list/table traversing.
If a table contains a reference to itself via a list inside the table
(or vice versa), this can cause a stack overflow when searching for
references with AdminShowReferences or when saving the game. Added
measures to prevent this from happening.

Also combined string marking in GC with the object renumbering for a
slight performance boost.