Convert nameid storage from linked list to tables, add admin.txt log.
The names/ids for message and parameter names were being kept track of
by a linked list, which was used infrequently but incurred a lookup time
of ~50 microseconds when getting an id or name. This was mostly used
during GC/savegame.

The nameids have been converted to a table system similar to the way
resources are stored by the server, which has a lookup time of <1
microsecond. Also added an admin show command to print out the nameids
(not very useful except for testing).

Moved the reload system calls from interface.c and adminfn.c to main.c,
to keep them in one place.

Created a new log file, admin.txt, which logs the output of aprintf()
(i.e. admin commands and the results). aprintf() moved to channel.c.

C_AddTableEntry now disallows keys to be made from objects, lists,
timers, tables or classes, as the numbers can change over time.
C_DeleteTable will log an error if called (and does not delete
anything).