Refactor client preference menu gameplay options.
The good copy of the preference menu options (safety, tempsafe,
grouping, reagentbag, autoloot, autocombine and spellpower display)
are now stored on the server in a new Player property, piPreferences.
These flags have been removed from piFlags, and the server/client
flags removed from object flags and placed in a new bitfield. This
is requested from the server when the client connects, or otherwise
requires fresh data from the server, and receiving the preferences
sets the client's preference menu.

Users can change preferences in the menu or with text commmands and the
preferences are all sent to the server. Any changed flags will cause
the familiar text responses to be sent to the client.

Players have all bits set to 1 on creation except for safety, which has
been renamed CF_SAFETY_OFF to reflect its status better (the flag being
on means safety is off).