Add GetListElemByClass and GetListNode C functions.
These C calls will be used to replace some constructs where a list is
iterated through to find an element not covered by FindListElem.

GetListElemByClass(list, class) takes a list and a class, and compares
the given class with each item in the list. Returns the list element
if found, otherwise returns $. Only compares objects.

GetListNode(list, position, element) returns the sublist of a list that
contains the element at the given position.