Move trestype TID prop to classvar, store trestypes as table in System.
Moved the TID constant from a property in each TreasureType to a
classvar, as this shouldn't be edited anyway. TreasureTypes are now
stored as a table in System, as it doesn't need to be iterated through
at any point - the only operation performed is lookup.

Changed GetSun to use GetListElemByClass rather than list iteration.