Refactor bank withdrawls/deposits, handle integer overflow.
Bank transactions were being treated too much like vault withdrawls -
the weight of shillings was checked when this is 0 (and has to be 0,
otherwise players couldn't hold a usable amount). This was also causing
an error if an immortal tried to withdraw money. This now checks
validity against MAX_KOD_INT instead, to prevent integer overflow (both
ways).