Add OwnerKilled placeholder for future use.
Called on an item if a player is killed while holding it. Passes the
message on to itematts.

Ao3 and SotH already had this message which was posted from
SomethingKilled() if the victim was the owner with the intention of
having the message run after Player.Killed() was finished. Calling it
from the item drop section of Killed() has the same effect.