Add Spectate spell, move spectator-specific code to new spell.
The Spectate effect is useful for admins to have, but requires custom
messages to be entered into admin window to activate. This change moves
the spectator-specific code from Player into a new DM spell, Spectate,
complete with new icon. The spell works the same as the spectator effect
did previously.

Survival arena applies the effect through this spell, and it is also
available to admins/DMs. Casting the spell on a player outside survival
arena requires the player to log off to remove the spell, as spectated
players cannot be cast on (DMs can remove the spell on themselves by
casting it again).

Added RemoveAllMinions message to Battler to deal with killing off all
the minions for a spectated player.

Monsters will check for non-interaction mode each state check, to
determine whether to drop aggro (IsInCannotInteractMode was optimised in
a previous commit and is now callable from such frequently called
messages). WentSpectatorMode removed from Room (due to added monster
checks).