Add ListCopy C call, fix list reallocation bugs.
Added a ListCopy() C call, takes a list and returns a copy of it.
Replaces the ListCopy() message in UtilityFunctions, as performing the
work directly in C is much faster.

Fixed up some issues in list.c where list node were being accessed after
a potential reallocation of the memory the list node points to. This
adds a little overhead to the list calls, but the alternative is
potentially crashing the server if the memory is reallocated.

Adjusted the timers on RecreateAll - reduced the time spent waiting
between stages to 500ms. Each stage executes far quicker than this, but
leaves room for slower computers and expansion.