Use reference vals for local vars when calling GetStepTowardsBSP.
Passes the local var ID for iRow, iCol, iFineRow and iFineCol, which are
then set directly in C_GetStepTowardsBSP instead of passing the values
back. C_GetStepTowardsBSP now just passes back state if the move was OK,
$ otherwise.

This include the blakcomp compiler code to enable passing IDs for local
vars.