Fix client not being able to move right to edge of things bbox.
The client could only move to the edge of the max row/cols, whereas now
it needs to be able to move to the fineunit edge. The client will now
load the things in the same manner as the server, and use them for room
boundary checking. The server will move the player to the next screen
when they hit the boundary.