Fix NPC random timer/spamlist behavior.
FirstUserEntered needs to check for AI_NPC when starting basic timers,
as NPCs are always in the wait state.

Added a check for no random timer to AddToSpamList, so that this timer
gets created if it isn't present. This will ensure the spam list gets
cleared when the timer goes off next.

Clear spam list when last user leaves the room (random timer is also
deleted here).