Fix for mobs (rarely) walking through walls.
When a monster's move will cross a wall, the point on the wall for the
lines move_start->move_end and the wall line itself is calculated.
Currently the X,Y values for this point are often incorrect by < 0.1,
and if the end point is across a wall from the start point and the end
point is very close to the wall, this will allow the monster to
incorrectly move through it.

To fix this, the calculation is now done using doubles instead of floats
(fixing cases where c1 is large). Some cases remained after this and to
handle those, the ISINBOX macro now subtracts 0.1f from the min box
value and adds 0.1f to the max box value to ensure the point is
contained within.