Unholy Touch health drain
Added a 1 point self-heal for wielders of Unholy Touch. It doesn't
actually 'drain' the target (your damage is doing that), but it does
heal you when you hit for 1 point, sometimes 2 (1/6th of the time).

Also added an IsDrainImmune check to monsters for future content. I
recommend Shadowbeasts and Undead be drain immune, but I didn't change
that here.