Refactor Loadout, fix to work with SpellItems.
Loadout will now handle SpellItems correctly - the number of charges on
each SpellItem in the player's inventory is recorded when Set is cast,
and Loadout will try to match these when stocking from the chest.

SpellItems are replenished from largest stack to smallest (according to
the loadout list). SpellItems are not combined when pulled into player's
inventory.