Upgrade list/search related admin commands.
- The "show matches" admin command will now return a list containing any
found objects.
- Added an admin "send list" command that will send a message to each
object in a list. This uses SendTopLevelBlakodMessage directly, and not
the SendList() C call (which shouldn't be used from outside KOD). This
admin function is also timed.
- The "show list" related commands now report the length of the list
found (and the length of any sublists).
- Converted the timing code in adminfn.c to use GetMicroCountDouble()
instead of QPC directly, as the former has a linux implementation.
- Add "create table" admin command which takes an integer (size) and
creates a table of that size.