Use object->num_props for bounds check in StoreValue.
To ensure that data isn't written outside an object's property array,
StoreValue must check lower (0) and upper (number of obj props) bounds
when writing data to the property.

Currently this is done using the number of properties taken from the
object's class data structure, requiring the class to be obtained and
checked to store the property. However, since objects already keep track
of how many properties they have, this value can be used directly from
the object itself, allowing the removal of the class-related code from
StoreValue.

This provides a 10-15% speed increase when assigning data to a property.