Add max list size and text buffer to 'show matches' cmd.
Show matches can potentially return tens or hundreds of thousands of
objects. To reduce the chances of this hanging the server, there is now
a max number of items show matches will add to a list (default 10000)
which can be changed via the MatchesList integer config option (set
config int [blakod] matcheslist #).

Also added a buffer for the strings being built for each matching
object, as most of the time in 'show matches' is spent printing to
log/terminal.