Remove server CR/LF replacement in logged strings.
All the channel-related functions check strings for the '\n' and convert
this to '\r\n', however this is only needed for displaying admin output
to the client.

All calls to TermConvertBuffer() in the server have been removed, and
the client will now process admin output with a slightly modified
function (AdminDlgConvertString()) before displaying to the user.

Refactored TraceInfo() to print output faster. Client admin window can
now display reasonable sized traces, though this isn't recommended as
the large number of packets can still crash the client (i.e. more than a
few hundred lines).

Changed the string buffer size on dprintf, eprintf etc. from 1000/2000
to BUFFER_SIZE (10000) to match max buffer size used elsewhere. The
channels themselves will only print a max string length of 1000, but the
full output will go to logs. Only aprintf currently handles strings of
this size (and is not printed to a channel); this is more of a safety
precaution.