Fix LoS issue with raised objects in small spaces.
Ivory Chapel guild hall chest is in a small raised alcove, and currently
BSPLineOfSightTree returns false when a player tries to add or remove
items from this chest (DEBUGLOS reports a ceiling block). This appears
to happen because of the height adjustment to the end point's Z axis,
which adjusts for the object's (which is presumed to be a player or
monster) viewing height.

BSPLineOfSight will now call BSPLineOfSightTree an extra time with end
point coordinates and height adjustment +1 (floor height returns a
block), which fixes this issue.

Also changed the chest LoS check in user.kod to check from player->chest
rather than chest->player.

Raised the time the Ivory Chapel chest stays open from 10 seconds to 60.